Fix kinematics unit test

+++ shared/utility/motion/InverseKinematics.cpp
@@ -413,9 +413,9 @@ namespace motion {
             return positions;
         }
 
-        std::vector<std::pair<ServoID, float>> calculateCameraLookJoints(const KinematicsModel& model,
-                                                                         const Eigen::Vector3d& cameraUnitVector) {
-            std::vector<std::pair<ServoID, float>> positions;
+        std::vector<std::pair<ServoID, double>> calculateCameraLookJoints(const KinematicsModel& model,
+                                                                          const Eigen::Vector3d& cameraUnitVector) {
+            std::vector<std::pair<ServoID, double>> positions;
             positions.push_back(
                 std::make_pair(ServoID::HEAD_YAW, std::atan2(cameraUnitVector.y(), cameraUnitVector.x())));
             positions.push_back(std::make_pair(ServoID::HEAD_PITCH,
